aiagent/requirements.txt
numpy
pandas
scikit-learn
fastapi
uvicorn
pymongo
python-dotenv
certifi
